+++ packages/ui-core/src/utils/cn.ts
@@ -1,8 +1,15 @@
 import classnames from 'classnames';
 
-interface ClassSet {
+interface IModificators {
+    [key: string]: boolean | string | undefined | null;
+}
+
+interface IClassSet {
     [key: string]: boolean;
 }
+
+type ElementNameType = string | IModificators | string[];
+type customClassName = string | string[];
 
 /**
  * Бэм генератор
@@ -14,20 +21,25 @@ interface ClassSet {
  * @returns {String}
  */
 export default function cnCreate(blockName: string) {
-    return (elementName: string, modificatorsObject?: {}, customClassNames?: string) => {
-        let params;
-        const prefix = elementName;
-        const className = customClassNames || '';
-
-        if (modificatorsObject) {
-            params = modificatorsObject;
-        }
+    function getElement(
+        elementName: ElementNameType,
+        modificatorsObject?: ElementNameType,
+        customClassNames?: customClassName
+    ) {
+        const prefix = typeof elementName === 'string' ? elementName : '';
+        const className = getCustomClassName(elementName, modificatorsObject, customClassNames);
+        const secondMods = !Array.isArray(modificatorsObject) && typeof modificatorsObject !== 'string'
+            ? modificatorsObject
+            : undefined;
+        const params = typeof elementName !== 'string' && !Array.isArray(elementName)
+            ? elementName
+            : secondMods;
 
         if (!params || typeof params === 'object' && Object.keys(params).length === 0) {
             return classnames(`${blockName}${prefix ? '__' + prefix : ''}`, className);
         }
 
-        const classParams: ClassSet = {};
+        const classParams: IClassSet = {};
         let prefixKey, withoutPrefix;
 
         for (const key in params) {
@@ -44,5 +56,27 @@ export default function cnCreate(blockName: string) {
         return prefix === ''
             ? classnames(blockName, classParams, className)
             : classnames(`${blockName}__${prefix}`, classParams, className);
-    };
+    }
+
+    return getElement;
 }
+
+const getCustomClassName = (first: ElementNameType, second?: ElementNameType, third?: customClassName) => {
+    if (Array.isArray(first)) {
+        return first.filter(f => typeof f === 'string').join(' ');
+    }
+    if (Array.isArray(second)) {
+        return second.filter(f => typeof f === 'string').join(' ');
+    }
+    if (typeof second === 'string') {
+        return second;
+    }
+    if (Array.isArray(third)) {
+        return third.filter(f => typeof f === 'string').join(' ');
+    }
+    if (typeof third === 'string') {
+        return third;
+    }
+
+    return '';
+};